# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3
info:
  title: API de Autenticação
  version: 1.0.0

tags:
  - name: Login
    description: Endpoints relacionados à autenticação

paths:
  /auth/login:
    get:
      tags:
        - Login
      operationId: AuthStatus
      summary: Consulta login por usuário
      parameters:
        - name: usuario
          in: query
          description: usuario
          required: true
          schema:
            type: string
            example: "fulano"
      responses:
        '200':
          description: Consulta realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatus'
    post:
      tags:
        - Login
      summary: Realizar login
      operationId: AuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogin'
      responses:
        '200':
          description: Login realizado com sucesso

  /auth/registration:
    post:
      tags:
        - Login
      summary: Criar login de usuário
      operationId: AuthRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegister'
      responses:
        '201':
          description: Login criado com sucesso

  /login/{usuario}:
    patch:
      tags:
        - Login
      summary: Atualizar senha de usuário
      operationId: UpdatePassword
      parameters:
        - name: usuario
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePassword'
      responses:
        '200':
          description: Senha atualizada com sucesso

    delete:
      tags:
        - Login
      summary: Deletar login de usuário
      operationId: DeleteLogin
      parameters:
        - name: usuario
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Login deletado com sucesso

components:
  schemas:
    AuthRegister:
      type: object
      required:
        - usuario
        - senha
        - documento
      properties:
        usuario:
          type: string
        senha:
          type: string
        documento:
          type: string

    AuthLogin:
      type: object
      required:
        - usuario
        - senha
      properties:
        usuario:
          type: string
        senha:
          type: string

    AuthStatus:
      type: object
      properties:
        status:
          type: string
          example: "ativo"

    UpdatePassword:
      type: object
      required:
        - senhaAntiga
        - senhaNova
      properties:
        senhaAntiga:
          type: string
        senhaNova:
          type: string